feat: ui scaffolding for limit orders pending design changes
