+++ src/components/MultiHopTrade/components/LimitOrder/types.ts
@@ -1,0 +1,5 @@
+export enum LimitOrderRoutePaths {
+  Input = '/trade/limit-order/input',
+  Confirm = '/trade/limit-order/confirm',
+  Status = '/trade/limit-order/status',
+}
