+++ src/components/MultiHopTrade/components/TradeInput/components/Claim/ClaimStatus.tsx
@@ -1,14 +1,12 @@
-import { CheckCircleIcon, WarningTwoIcon } from '@chakra-ui/icons'
-import { Button, CardBody, CardFooter, Center, Heading, Link, Stack } from '@chakra-ui/react'
+import { Button, CardBody, CardFooter, Link } from '@chakra-ui/react'
 import { TxStatus } from '@shapeshiftoss/unchained-client'
 import { AnimatePresence } from 'framer-motion'
 import React, { useCallback, useMemo } from 'react'
 import { useTranslate } from 'react-polyglot'
 import { useHistory } from 'react-router'
 import { Amount } from 'components/Amount/Amount'
-import { CircularProgress } from 'components/CircularProgress/CircularProgress'
+import { StatusBody } from 'components/MultiHopTrade/components/StatusBody'
 import { SlideTransition } from 'components/SlideTransition'
-import { SlideTransitionY } from 'components/SlideTransitionY'
 import { useSafeTxQuery } from 'hooks/queries/useSafeTx'
 import { getTxLink } from 'lib/getTxLink'
 import { fromBaseUnit } from 'lib/math'
@@ -17,43 +15,6 @@ import { useAppSelector } from 'state/store'
 
 import type { ClaimDetails } from './hooks/useArbitrumClaimsByStatus'
 import { ClaimRoutePaths } from './types'
-
-type ClaimStatusBodyProps = {
-  amountCryptoPrecision: string
-  description: string
-  icon: JSX.Element
-  status: TxStatus
-  symbol: string
-  title: string
-}
-
-export const ClaimStatusBody: React.FC<ClaimStatusBodyProps> = ({
-  amountCryptoPrecision,
-  description,
-  icon,
-  status,
-  symbol,
-  title,
-}) => {
-  return (
-    <SlideTransitionY key={status}>
-      <CardBody py={32}>
-        <Center flexDir='column' gap={4}>
-          {icon}
-          <Stack spacing={2} alignItems='center'>
-            <Heading as='h4'>{title}</Heading>
-            <Amount.Crypto
-              prefix={description}
-              value={amountCryptoPrecision}
-              symbol={symbol}
-              color='text.subtle'
-            />
-          </Stack>
-        </Center>
-      </CardBody>
-    </SlideTransitionY>
-  )
-}
 
 type ClaimStatusProps = {
   activeClaim: ClaimDetails
@@ -79,53 +40,47 @@ export const ClaimStatus: React.FC<ClaimStatusProps> = ({
     return fromBaseUnit(activeClaim.amountCryptoBaseUnit, asset?.precision ?? 0)
   }, [asset, activeClaim])
 
-  const StatusBody = useMemo(() => {
+  const statusBody = useMemo(() => {
     if (!asset) return null
+    const renderedAmount = (() => {
+      switch (claimTxStatus) {
+        case TxStatus.Pending: {
+          return (
+            <Amount.Crypto
+              prefix={translate('bridge.claimPending')}
+              value={amountCryptoPrecision}
+              symbol={asset.symbol}
+              color='text.subtle'
+            />
+          )
+        }
+        case TxStatus.Confirmed: {
+          return (
+            <Amount.Crypto
+              prefix={translate('bridge.claimSuccess')}
+              value={amountCryptoPrecision}
+              symbol={asset.symbol}
+              color='text.subtle'
+            />
+          )
+        }
+        case TxStatus.Failed: {
+          return (
+            <Amount.Crypto
+              prefix={translate('bridge.claimFailed')}
+              value={amountCryptoPrecision}
+              symbol={asset.symbol}
+              color='text.subtle'
+            />
+          )
+        }
+        case TxStatus.Unknown:
+        default:
+          return null
+      }
+    })()
 
-    switch (claimTxStatus) {
-      case TxStatus.Pending: {
-        return (
-          <ClaimStatusBody
-            amountCryptoPrecision={amountCryptoPrecision}
-            description={translate('bridge.claimPending')}
-            // eslint-disable-next-line react-memo/require-usememo
-            icon={<CircularProgress size='24' />}
-            status={TxStatus.Pending}
-            symbol={asset.symbol}
-            title={translate('pools.waitingForConfirmation')}
-          />
-        )
-      }
-      case TxStatus.Confirmed: {
-        return (
-          <ClaimStatusBody
-            amountCryptoPrecision={amountCryptoPrecision}
-            description={translate('bridge.claimSuccess')}
-            // eslint-disable-next-line react-memo/require-usememo
-            icon={<CheckCircleIcon color='text.success' boxSize='24' />}
-            status={TxStatus.Confirmed}
-            symbol={asset.symbol}
-            title={translate('common.success')}
-          />
-        )
-      }
-      case TxStatus.Failed: {
-        return (
-          <ClaimStatusBody
-            amountCryptoPrecision={amountCryptoPrecision}
-            description={translate('bridge.claimFailed')}
-            // eslint-disable-next-line react-memo/require-usememo
-            icon={<WarningTwoIcon color='red.500' boxSize='24' />}
-            status={TxStatus.Failed}
-            symbol={asset.symbol}
-            title={translate('common.somethingWentWrong')}
-          />
-        )
-      }
-      case TxStatus.Unknown:
-      default:
-        return null
-    }
+    return <StatusBody txStatus={claimTxStatus}>{renderedAmount}</StatusBody>
   }, [amountCryptoPrecision, asset, translate, claimTxStatus])
 
   const { data: maybeSafeTx } = useSafeTxQuery({
@@ -144,7 +99,9 @@ export const ClaimStatus: React.FC<ClaimStatusProps> = ({
 
   return (
     <SlideTransition>
-      <AnimatePresence mode='wait'>{StatusBody}</AnimatePresence>
+      <AnimatePresence mode='wait'>
+        <CardBody py={32}>{statusBody}</CardBody>
+      </AnimatePresence>
       <CardFooter flexDir='column' gap={2}>
         <Button as={Link} href={txLink} size='lg' colorScheme='blue' variant='ghost' isExternal>
           {translate('trade.viewTransaction')}
